speech_service/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
ffmpeg-python
